fix checking of existing bridges
The comment states "Check to see if >any< bridges exist[...]", while
it actually checks for the hard-coded bridges "bridge0" and "bridge1".
This causes iocell to always create those two bridges if present bridges
on the system have other names.
Because the default route interface is being added to 'bridge0', the
interface resets and drops all connections - on a jailhost this might be
the default mgmt interface which carries the connection over which
iocell was invoked and thus being terminated in the middle of starting
the jail.

This Fix changes the grep'ing for bridge0|bridge1 to simply checking for
interfaces belonging to the 'bridge' group via 'ifconfig -g bridge'.