Add support for vnet.interface parameter via vnet_interfaces property
This adds support for the vnet.interface jail parameter, which allows
to directly pass interfaces (mainly SR-IOV VFs) to a jail.
The vnet_interfaces property can be set to a single interface or a
comma-separated list of interfaces; which are passed as vnet.interface
parameters to jail(8).
The default is "none"